requirements.txt
numpy==1.18.1
matplotlib==3.1.2
scipy==1.4.1
Pillow==3.1